solved the issue for user stay logged in
