+++ src/features/account/components/form-forgot-password.components.js
@@ -30,7 +30,7 @@ const SendButtonSection = styled.View`
 
 const SendText = styled(Text)`
   color: white;
-  font-family: ${(props) => props.theme.fonts.body_medium};
+  font-family: ${(props) => props.theme.fonts.body_400};
   font-size: ${(props) => props.theme.fontSizes.body};
 `;
