+++ src/features/account/components/form-register.components.js
@@ -32,7 +32,7 @@ const SignUpButtonSection = styled.View`
 
 const SignUpText = styled(Text)`
   color: white;
-  font-family: ${(props) => props.theme.fonts.body_medium};
+  font-family: ${(props) => props.theme.fonts.body_400};
   font-size: ${(props) => props.theme.fontSizes.button};
 `;
