+++ src/features/account/screens/forgot-password.screen.js
@@ -16,12 +16,12 @@ const TopTextSection = styled.View`
 `;
 
 const ForgotPasswordText = styled(Text)`
-  font-family: ${(props) => props.theme.fonts.body_bold};
+  font-family: ${(props) => props.theme.fonts.body_600};
   font-size: ${(props) => props.theme.fontSizes.title};
 `;
 
 const MessageText = styled(Text)`
-  font-family: ${(props) => props.theme.fonts.body_bold};
+  font-family: ${(props) => props.theme.fonts.body_600};
   font-size: ${(props) => props.theme.fontSizes.body};
   text-align: center;
 `;
@@ -40,7 +40,7 @@ const OrSection = styled.View`
 `;
 
 const OrText = styled(Text)`
-  font-family: ${(props) => props.theme.fonts.body_medium};
+  font-family: ${(props) => props.theme.fonts.body_400};
   font-size: ${(props) => props.theme.fontSizes.body};
   text-align: center;
   align-self: center;
