+++ src/features/account/screens/register.screen.js
@@ -20,7 +20,7 @@ const SignInSection = styled.View`
 `;
 
 const CreateAccountText = styled(Text)`
-  font-family: ${(props) => props.theme.fonts.body_bold};
+  font-family: ${(props) => props.theme.fonts.body_600};
   font-size: ${(props) => props.theme.fontSizes.h4};
 `;
 
@@ -47,7 +47,7 @@ const OrSection = styled.View`
 `;
 
 const OrText = styled(Text)`
-  font-family: ${(props) => props.theme.fonts.body_medium};
+  font-family: ${(props) => props.theme.fonts.body_400};
   font-size: ${(props) => props.theme.fontSizes.body};
   text-align: center;
   align-self: center;
