+++ src/infrastructure/navigation/index.js
@@ -9,12 +9,9 @@ import { AuthenticationContext } from "../../services/authentication/authenticat
 export const Navigation = () => {
   const { isAuthenticated } = useContext(AuthenticationContext);
 
-  // console.log(isAuthenticated);
-
   return (
     <NavigationContainer>
-      {/* {isAuthenticated ? <AppNavigator /> : <AccountNavigator />} */}
-      <AppNavigator />
+      {isAuthenticated ? <AppNavigator /> : <AccountNavigator />}
     </NavigationContainer>
   );
 };
